backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # Not needed for SQLite
PyJWT==2.10.1
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
openai==1.3.0
stripe==7.6.0
celery==5.3.4
redis==5.0.1
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2
Pillow==10.1.0
boto3==1.34.0
httpx==0.25.2